+++ nikiforov.alexander/T2/iofmtguard.hpp
@@ -1,0 +1,20 @@
+#ifndef IOFMTGUARD_HPP
+#define IOFMTGUARD_HPP
+#include <sstream>
+
+namespace nikiforov
+{
+  class iofmtguard
+  {
+  public:
+    explicit iofmtguard(std::basic_ios< char >& s) noexcept;
+    ~iofmtguard();
+  private:
+    std::basic_ios< char >& s_;
+    char fill_;
+    std::streamsize precision_;
+    std::basic_ios< char >::fmtflags fmt_;
+  };
+}
+
+#endif
