+++ nikiforov.alexander/T2/typesKeys.hpp
@@ -1,0 +1,33 @@
+#ifndef TYPESKEYS_HPP
+#define TYPESKEYS_HPP
+#include <sstream>
+
+namespace nikiforov
+{
+  struct DelimiterIO
+  {
+    char exp;
+  };
+
+  struct UnsignedLongLongIO
+  {
+    unsigned long long& ref;
+  };
+
+  struct CharIO
+  {
+    char& ref;
+  };
+
+  struct StringIO
+  {
+    std::string& ref;
+  };
+
+  std::istream& operator>>(std::istream& in, DelimiterIO&& dest);
+  std::istream& operator>>(std::istream& in, UnsignedLongLongIO&& dest);
+  std::istream& operator>>(std::istream& in, CharIO&& dest);
+  std::istream& operator>>(std::istream& in, StringIO&& dest);
+}
+
+#endif
